\name{as.climdex.data.frame}
\alias{as.climdex.data.frame}
\title{Coercion to a ClimDex Data Frame}
\usage{
  as.climdex.data.frame(data)
}
\arguments{
  \item{data}{the object to be transformed}
}
\description{
  This functions transforms a generic data object
  \code{data} in a \code{clindex.data.frame}-type S3 object
}
\author{
  Emanuele Cordano, Annalisa Di Piazza
}
\seealso{
  \code{\link{climdex.data.frame}}

  \code{\link{climdex.data.frame}}
}

