\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.climdex.data.frame}
\title{Trasformation of a ClimDex Data Frame to a Data Frame}
\usage{
  \method{as.data.frame}{climdex.data.frame} (x, ...)
}
\arguments{
  \item{x}{the object to be transformed}

  \item{...}{further arguments}
}
\description{
  This method transforms a \code{climdex.data.frame}-type
  S3 object into a \code{data.frame} object
}
\author{
  Emanuele Cordano, Annalisa Di Piazza

  Emanuele Cordano, Annalisa Di Piazza
}
\seealso{
  \code{\link{as.climdex.data.frame}},\code{\link{as.data.frame}}
}

