\name{temperature_max_daily}
\alias{temperature_max_daily}
\title{Daily Maximum Temperature}
\usage{
  temperature_max_daily(x)
}
\arguments{
  \item{x}{an object of class
  \code{\link{climdexInput-class}}}
}
\value{
  Daily Maximum Temperature
}
\description{
  Extracts daily maximum temparature from an object of
  class \code{\link{climdexInput-class}}.
}
\author{
  Emanuele Cordano, Annalisa Di Piazza
}
\seealso{
  \code{\link{climdexInput-class}},
  \code{\link{climdexInput.raw}}
}

