\name{temperature_min_daily}
\alias{temperature_min_daily}
\title{Daily Minimum Temperature}
\usage{
  temperature_min_daily(x)
}
\arguments{
  \item{x}{an object of class
  \code{\link{climdexInput-class}}}
}
\value{
  Daily Minimum Temperature
}
\description{
  Extracts daily Minimum temparature from an object of
  class \code{\link{climdexInput-class}}.
}
\author{
  Emanuele Cordano, Annalisa Di Piazza
}
\seealso{
  \code{\link{climdexInput-class}},
  \code{\link{climdexInput.raw}}
}

